Extract HasLocallyFree out of implicits package object
